publisher/requirements.txt
paho-mqtt==2.0.0
python-dotenv==1.0.1